backend/requirements.txt
amqp==5.1.1
asgiref==3.6.0
auth0-python==4.1.0
Authlib==1.0.1
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cryptography==40.0.1
Django==4.0.10
django-cors-headers==3.14.0
dnspython==2.3.0
eventlet==0.33.3
greenlet==2.0.2
idna==3.4
kombu==5.2.4
prompt-toolkit==3.0.38
pycparser==2.21
PyJWT==2.6.0
python-dotenv==0.19.2
pytz==2023.3
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.8
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
